Fix a bug of several ModelComponents being added to the same SceneActor twice.
This minor bug affects several demo projects where a ModelComponent is added to a SceneActor using CreateAndAddComponent<T>(), and later an attempt is made to add it again using AddComponent().

In practice, it doesn’t impact the demos, since we explicitly check to ensure that the same type of Component can only be added to a SceneActor once.